Refactored BootChargingStationEvent to support attributes
Note: this is a step in a series of refactorings which will lead to lower coupling between Motown and OCPP. For now, the assumption with which this event is refactored is that only the data required to build the logical model in the Motown domain application is set as a field in the event objects. All other information is added as a Map so applications which are built on top of the Motown domain application can use this as they wish.