Added configured and unconfigured boot events
The ChargingStationBootedEvent is now abstract and has two subclasses,
the UnconfiguredChargingStationBootedEvent and the
ConfiguredChargingStationBootedEvent. Event listeners only interested in
the fact that a charging station has booted can listen to the abstract
super class while listeners interested in only one of the specific
variations of a ChargingStationBootedEvent can listen to one of the
subclasses.