Removed unused status type from Core API
The ChargingStationRegistrationStatus type was initially used to
directly communicate results back to the caller when the OCPP addon did
not have a view model yet. As the OCPP addon now has a view model and
doesn't use this return value it has been removed to keep the Core API
to a minimum.