Removed unused constructor from CorrelationToken
The CorrelationToken constructor with a string parameter was only used in tests and these tests were easily adapted to the new situation. Removing the constructor also gives us the guarantee that CorrelationTokens are UUIDs which makes them more flexible for our use.

For instance, we can now use them as (guaranteed to be unique) message IDs in the OCPP/J implementation. This gives us the added advantage that we can trace commands to events to messages to the charging station.