Add DataTransferMessage type
In an attempt to lower code duplication and unit interfacing this commit
introduces the DataTransferMessage type which bundles the vendor
identifier, message identifier, and data for data transfer related
commands and events. This commit only introduces the type and uses it in
the IncomingDataTransferCommand and all the code interacting with this
command.