Selective identification authorization service
This service can be used to specify authorization providers for each
charging station. The module contains a charging station entity
which has a field `authorizationProviders` which should contain a
comma-separated list of authorization providers to be used for that
charging station. The identifier of the authorization provider is
configured in the Spring context in the
`SelectiveIdentificationAuthorizationService`. An example is 'ocpi' or
'cir'.

#66